refactor: jwt_handler added and tests and codes refactored

+++ src/services/user_service/api/user_router.py
@@ -1,14 +1,16 @@
 from ..services.user_service import UserService
 from ..models.user_model import User
-from ..schemas.api_schemas import (UserCreate, UserCreatedSuccesfully, UserGet, UserUpdate, UserUpdated,
-                                                       UserDeleted, UserActive, UserAuth, UserAuthed)
+from ..schemas.api_schemas import (UserCreate, UserCreatedSuccesfully, UserGet, UserUpdate,
+                                   UserUpdated, UserActive, UserAuth, UserAuthed)
 from ..database.mongo import user_collection
 from ..repositories.user_repository_mongodb import MongoUserRepository
 from fastapi import APIRouter, HTTPException, Depends
+from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from ..utils.jwt_handler import get_current_user
 import logging
 
 logger = logging.getLogger(__name__)
-
+security = HTTPBearer()
 
 class UserRouter:
     def __init__(self):
@@ -18,6 +20,7 @@ class UserRouter:
     def get_user_service(self):
         repo = MongoUserRepository(user_collection)
         return UserService(repo)
+    
     def register_routes(self):
         @self.router.post("/")
         async def create_user(user: UserCreate, service: UserService = Depends(self.get_user_service))-> UserCreatedSuccesfully:
@@ -25,40 +28,65 @@ class UserRouter:
                 user = User(**user.model_dump())
                 logger.info(f"Create user request received for email={user.email}")
                 return await service.add_user(user)
+            
             except ValueError as e:
                 raise HTTPException(status_code=409, detail=str(e))
             
         @self.router.get("/{user_id}")
-        async def get_user(user_id: str, service: UserService = Depends(self.get_user_service))-> UserGet:
+        async def get_user(user_id: str, service: UserService = Depends(self.get_user_service),
+                              payload: dict = Depends(get_current_user))-> UserGet:
+            
+            if payload.get("id") != user_id:
+                raise HTTPException(status_code=403, detail="Forbidden")
+            
             try:
                 logger.info(f"Get user request received for user_id={user_id}")
                 return await service.get_user(user_id)
+            
             except ValueError as e:
                 raise HTTPException(status_code=404, detail=str(e))
             
         @self.router.put("/{user_id}")
-        async def update_user(user_id: str, data: UserUpdate, service: UserService = Depends(self.get_user_service))-> UserUpdated:
+        async def update_user(user_id: str, data: UserUpdate, service: UserService = Depends(self.get_user_service),
+                              payload: dict = Depends(get_current_user))-> UserUpdated:
+            
+            if payload.get("id") != user_id:
+                raise HTTPException(status_code=403, detail="Forbidden")
+            
             try:
                 logger.info(f"Update user request received for user_id={user_id}")
                 return await service.update_user(user_id,data.model_dump(exclude_unset=True))
+            
             except ValueError as e:
                 raise HTTPException(status_code=404, detail=str(e))
             except FileExistsError as e:
                 raise HTTPException(status_code=409, detail=str(e))
             
         @self.router.delete("/{user_id}", status_code=204)
-        async def delete_user(user_id: str, service: UserService = Depends(self.get_user_service)):
+        async def delete_user(user_id: str, service: UserService = Depends(self.get_user_service),
+                              payload: dict = Depends(get_current_user)):
+            
+            if payload.get("id") != user_id:
+                raise HTTPException(status_code=403, detail="Forbidden")
+            
             try:
                 logger.info(f"Delete user request received for user_id={user_id}")
                 await service.delete_user(user_id)
+                
             except ValueError as e:
                 raise HTTPException(status_code=404, detail=str(e))
             
         @self.router.get("/{user_id}/active")
-        async def is_user_active(user_id: str, service: UserService = Depends(self.get_user_service))-> UserActive:
+        async def is_user_active(user_id: str, service: UserService = Depends(self.get_user_service),
+                              payload: dict = Depends(get_current_user))-> UserActive:
+            
+            if payload.get("id") != user_id:
+                raise HTTPException(status_code=403, detail="Forbidden")
+            
             try:
                 logger.info(f"Is user active request received for user_id={user_id}")
                 return await service.get_user(user_id)
+            
             except ValueError as e:
                 raise HTTPException(status_code=404, detail=str(e))
             
@@ -66,8 +94,10 @@ class UserRouter:
         async def authenticate_user(data: UserAuth, service: UserService = Depends(self.get_user_service))-> UserAuthed:
             logger.info(f"Authenticate request received for email={data.email}")
             user = await service.authenticate_user(data.email,data.password)
+            
             if not user:
                 logger.warning(f"Authentication failed for email={data.email}")
                 raise HTTPException(status_code=401, detail="Invalid email or password")
+            
             logger.info(f"Authentication successful for user_id={user.id}")
             return user
